refactor: [#589] versioning for configuration
