+++ src/config/v1/image_cache.rs
@@ -1,0 +1,37 @@
+use serde::{Deserialize, Serialize};
+
+/// Configuration for the image proxy cache.
+///
+/// Users have a cache quota per period. For example: 100MB per day.
+/// When users are navigating the site, they will be downloading images that are
+/// embedded in the torrent description. These images will be cached in the
+/// proxy. The proxy will not download new images if the user has reached the
+/// quota.
+#[allow(clippy::module_name_repetitions)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct ImageCache {
+    /// Maximum time in seconds to wait for downloading the image form the original source.
+    pub max_request_timeout_ms: u64,
+    /// Cache size in bytes.
+    pub capacity: usize,
+    /// Maximum size in bytes for a single image.
+    pub entry_size_limit: usize,
+    /// Users have a cache quota per period. For example: 100MB per day.
+    /// This is the period in seconds (1 day in seconds).
+    pub user_quota_period_seconds: u64,
+    /// Users have a cache quota per period. For example: 100MB per day.
+    /// This is the maximum size in bytes (100MB in bytes).    
+    pub user_quota_bytes: usize,
+}
+
+impl Default for ImageCache {
+    fn default() -> Self {
+        Self {
+            max_request_timeout_ms: 1000,
+            capacity: 128_000_000,
+            entry_size_limit: 4_000_000,
+            user_quota_period_seconds: 3600,
+            user_quota_bytes: 64_000_000,
+        }
+    }
+}
